socperf功能失效修改

+++ services/core/src/socperf_config.cpp
@@ -676,7 +676,7 @@ bool SocPerfConfig::LoadResourceAvailable(std::shared_ptr<ResNode> resNode, cons
     std::vector<std::string> result = Split(nodeStr, SPLIT_SPACE);
     for (auto str : result) {
         if (IsNumber(str)) {
-            resNode->available.insert(atoi(str.c_str()));
+            resNode->available.insert(atoll(str.c_str()));
         } else {
             return false;
         }
